docker_players_api_sql/postgrestest/requirements.txt
flask
flask-restful
pymongo
python-dotenv
boto3
sqlalchemy
psycopg2-binary
# psycopg2
# sqlalchemy.ext.declarative
# sqlalchemy.orm